Add login/logout Navbar
