+++ client/src/components/Login/SignupUser-button.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import SignupButton from "../../components/Login/Signup-button"
+import AuthNav from "../../components/Login/Auth-nav";
+
+const SignupUserButton = () => {
+    return (
+        <div>
+            <SignupButton />
+            <AuthNav />
+        </div>
+    )
+}
+
+export default SignupUserButton
