+++ client/src/views/Landing/landing.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { useTheme } from '@material-ui/core/styles';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
-import AuthNav from "../../components/Login/Auth-nav";
-import SignupButton from "../../components/Login/Signup-button"
+// import AuthNav from "../../components/Login/Auth-nav";
+// import SignupButton from "../../components/Login/Signup-button"
+import AuthSignup from "../../components/Login/Auth-signup"
 
 import { Box, Container } from "@material-ui/core"
 import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
@@ -39,8 +40,9 @@ const Landing: React.FC = (): JSX.Element => {
             <Container >
                 <img className={classes.logo} src={matches ? "/fullLogo.png" : "/logo512.png"} alt="logo" />
             </Container>
-            <SignupButton />
-            <AuthNav />
+            {/* <SignupButton />
+            <AuthNav /> */}
+            <AuthSignup />
         </Box>
     )
 }
